+++ tortoise/sim/votes.go
@@ -23,30 +23,3 @@ func perfectVoting(rng *rand.Rand, layers []*types.Layer) Voting {
 	base := baseLayer.Blocks()[rng.Intn(len(baseLayer.Blocks()))]
 	return Voting{Base: base.ID(), Support: support}
 }
-
-// GapVote will skip one layer in voting.
-func GapVote(rng *rand.Rand, layers []*types.Layer) Voting {
-	if len(layers) == 1 {
-		return perfectVoting(rng, layers)
-	}
-	baseLayer := layers[len(layers)-1]
-	support := layers[len(layers)-2].BlocksIDs()
-	base := baseLayer.Blocks()[rng.Intn(len(baseLayer.Blocks()))]
-	return Voting{Base: base.ID(), Support: support}
-}
-
-// OlderExceptions will vote for block older then base block.
-func OlderExceptions(rng *rand.Rand, layers []*types.Layer) Voting {
-	if len(layers) == 1 {
-		return perfectVoting(rng, layers)
-	}
-	baseLayer := layers[len(layers)-1]
-	base := baseLayer.Blocks()[rng.Intn(len(baseLayer.Blocks()))]
-	voting := Voting{Base: base.ID()}
-	for _, layer := range layers[len(layers)-2:] {
-		for _, bid := range layer.BlocksIDs() {
-			voting.Support = append(voting.Support, bid)
-		}
-	}
-	return voting
-}
